Add -Wl,--start-group & -Wl,--end-group
It's probably a better way to add libraries when we link objects from
OCaml. With these options, we allow the linker to resolve all symbols
regardless the order of these libraries. Indeed, the linker will
repeatly lookup on these libraries until all symbols are resolved.

However, it seems that these options can have an impact about
performances of the linker. If we need to improve this situation, we
must recalculate a topological order of libraries. But for our purpose,
that's probably fine.

Co-authored-by: Pierre Alain <pierre.alain@tuta.io>
Co-authored-by: Kate <kit.ty.kate@disroot.org>
Co-authored-by: Christiano Haesbaert <haesbaert@haesbaert.org>